+++ sdk/src/integrationTest/java/FileDeleteIntegrationTest.java
@@ -9,12 +9,9 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Objects;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.assertj.core.api.Assertions.as;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 public class FileDeleteIntegrationTest {
     @Test
@@ -33,12 +30,12 @@ public class FileDeleteIntegrationTest {
             .setFileId(fileId)
             .execute(testEnv.client);
 
-        assertEquals(fileId, info.fileId);
-        assertEquals(28, info.size);
-        assertFalse(info.isDeleted);
-        assertNotNull(info.keys);
-        assertNull(info.keys.getThreshold());
-        assertEquals(KeyList.of(testEnv.operatorKey), info.keys);
+        assertThat(info.fileId).isEqualTo(fileId);
+        assertThat(info.size).isEqualTo(28);
+        assertThat(info.isDeleted).isFalse();
+        assertThat(info.keys).isNotNull();
+        assertThat(info.keys.getThreshold()).isNull();
+        assertThat(info.keys).isEqualTo(KeyList.of(testEnv.operatorKey));
 
         new FileDeleteTransaction()
             .setFileId(fileId)
@@ -63,19 +60,17 @@ public class FileDeleteIntegrationTest {
             .setFileId(fileId)
             .execute(testEnv.client);
 
-        assertEquals(fileId, info.fileId);
-        assertEquals(28, info.size);
-        assertFalse(info.isDeleted);
-        assertNull(info.keys);
+        assertThat(info.fileId).isEqualTo(fileId);
+        assertThat(info.size).isEqualTo(28);
+        assertThat(info.isDeleted).isFalse();
+        assertThat(info.keys).isNull();
 
-        var error = assertThrows(ReceiptStatusException.class, () -> {
+        assertThatExceptionOfType(ReceiptStatusException.class).isThrownBy(() -> {
             new FileDeleteTransaction()
                 .setFileId(fileId)
                 .execute(testEnv.client)
                 .getReceipt(testEnv.client);
-        });
-
-        assertTrue(error.getMessage().contains(Status.UNAUTHORIZED.toString()));
+        }).withMessageContaining(Status.UNAUTHORIZED.toString());
 
         testEnv.close();
     }