+++ sdk/src/integrationTest/java/LiveHashDeleteIntegrationTest.java
@@ -10,8 +10,7 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Objects;
 
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 class LiveHashDeleteIntegrationTest {
     private static final byte[] HASH = Hex.decode("100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000002");
@@ -30,15 +29,13 @@ class LiveHashDeleteIntegrationTest {
 
         var accountId = Objects.requireNonNull(response.getReceipt(testEnv.client).accountId);
 
-        var error = assertThrows(PrecheckStatusException.class, () -> {
+        assertThatExceptionOfType(PrecheckStatusException.class).isThrownBy(() -> {
             new LiveHashDeleteTransaction()
                 .setAccountId(accountId)
                 .setHash(HASH)
                 .execute(testEnv.client)
                 .getReceipt(testEnv.client);
-        });
-
-        assertTrue(error.getMessage().contains(Status.NOT_SUPPORTED.toString()));
+        }).withMessageContaining(Status.NOT_SUPPORTED.toString());
 
         testEnv.close(accountId, key);
     }