+++ sdk/src/main/java/com/hedera/hashgraph/sdk/Transaction.java
@@ -253,6 +253,9 @@ public abstract class Transaction<T extends Transaction<T>>
             case CONTRACTDELETEINSTANCE:
                 return new ContractDeleteTransaction(txs);
 
+            case ETHEREUMTRANSACTION:
+                return new EthereumTransaction(txs);
+
             case CRYPTOADDLIVEHASH:
                 return new LiveHashAddTransaction(txs);