+++ sdk/src/main/java/com/hedera/hashgraph/sdk/TransactionResponse.java
@@ -35,6 +35,7 @@ public final class TransactionResponse implements WithGetReceipt, WithGetRecord 
     public final TransactionId transactionId;
 
     @Nullable
+    @Deprecated
     public final TransactionId scheduledTransactionId;
 
     TransactionResponse(
@@ -62,6 +63,13 @@ public final class TransactionResponse implements WithGetReceipt, WithGetRecord 
         return receipt;
     }
 
+    public TransactionReceipt forceGetReceipt(Client client) throws TimeoutException, PrecheckStatusException {
+        return new TransactionReceiptQuery()
+            .setTransactionId(transactionId)
+            .setNodeAccountIds(Collections.singletonList(nodeId))
+            .execute(client);
+    }
+
     @Override
     public CompletableFuture<TransactionReceipt> getReceiptAsync(Client client) {
         return new TransactionReceiptQuery()
@@ -72,6 +80,14 @@ public final class TransactionResponse implements WithGetReceipt, WithGetRecord 
 
     public TransactionRecord getRecord(Client client) throws TimeoutException, PrecheckStatusException, ReceiptStatusException {
         getReceipt(client);
+
+        return new TransactionRecordQuery()
+            .setTransactionId(transactionId)
+            .setNodeAccountIds(Collections.singletonList(nodeId))
+            .execute(client);
+    }
+
+    public TransactionRecord forceGetRecord(Client client) throws TimeoutException, PrecheckStatusException {
 
         return new TransactionRecordQuery()
             .setTransactionId(transactionId)